app/pyproject.toml
[build-system]
requires = ["setuptools>=82", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PodInspector"
version = "1.0.0"
description = "PodInspector application for k8s."
authors = [
  { name="HawerPL" }
]

requires-python = ">=3.14"
dependencies = [
    "fastapi~=0.129.0",
    "pydantic~=2.12.5",
    "kubernetes~=35",
    "uvicorn~=0.40",
    "prometheus-fastapi-instrumentator~=7.1.0",
    "pyroscope-io~=1.0.1"
]

[tool.setuptools]
packages = ["api", "core", "services"]